slizzy chat/requirements.txt
Flask==3.1.0
spacy==3.8.2
Flask-Cors==5.0.0
gunicorn==23.0.0
openai==0.28.0
waitress==3.0.2
python-dotenv==1.0.1
python==3.12.6